Refactor weight calculation code in apply_temporal_filter()
In apply_temporal_filter(), few of the variables related to
weight calculation were calculated for every pixel though
they are constant at sub-block level. In this CL, the
relevant calculations are done beforehand and the same are
consumed based on the sub-block index in both SSE2 and AVX2
variants of the function.

For 'good' encoding,
               Instruction Count        BD-Rate Loss(%)
cpu Resolution   Reduction(%)    avg.psnr  ovr.psnr   ssim
 5   LOWRES2       0.201         0.0000     0.0000    0.0000
 5   MIDRES2       0.263         0.0000     0.0000    0.0000
 5    HDRES2       0.321         0.0000     0.0000    0.0000
 6   LOWRES2       0.115         0.0000     0.0000    0.0000
 6   MIDRES2       0.243         0.0000     0.0000    0.0000
 6    HDRES2       0.296         0.0000     0.0000    0.0000

Change-Id: I486418ff972940ac5863856a7364eb5374ed2b38