Adjust how fast Q can fall in cyclic refresh.
This change is targeted at screen content which often
is mainly static much of the time.

In a static scene the baseline Q gradually falls but we also
upgrade a small number of blocks each frame to a lower
q using cyclic refresh if the scene has low motion.

This change tracks the progress of the refresh cycle. at the start of
a new cycle through the frame Q is allowed to drop by a larger
step, but during the cycle by a smaller step. The idea here is that
by the end of a cycle most blocks will already have been updated
at a Q well below the baseline frame Q so a larger step down in
this value will be less likely to cause a big rate spike. However,
during a cycle we reduce the maximum per frame drop in Q
so that all the blocks have the opportunity to be "refreshed"

This change improved the metrics for the screen set as follows
for speed 10 (-ve = better) and also a small improvement in
encoding speed of 0.876%

OPSNR -1.191%, SSIM -1.291%

STATS_CHANGED

Change-Id: I9301e93e645a135d6e9a5e60af90834cc8d2d51d