Add accumulator overflow test for masked SAD functions
Accumulator overflow is unlikely to be caught by purely random testing.
This is due to the fact that a sum of uniformly distributed random
numbers will only hit the max/min possible sum with exponentially
reducing likelihood as the problem size grows.

The maximum accumulator value for an 8-bit masked SAD occurs when doing
an absolute difference of 255 and 0. To ensure that we correctly handle
large accumulator values, this commit adds an explicit test case with
such values.

Change-Id: Ida4544e908d8416ebd9341abb8abe7111186ec3c