Overhaul disflow algorithm
This patch applies a huge amount of bugfixes and optimizations to the
disflow algorithm in the flow estimation library. The new version of
this code generates better results than the current corner matching
code, while being many times faster.

This patch also includes an SSE4.1 implementation of aom_compute_flow_at_point,
which contains the core loop of the algorithm, for an additional speed boost
on x86.

This code is not enabled yet, so this patch does not directly affect BDRATE.
But preliminary results suggest that disflow with SSE4.1 gives a 2 - 2.5%
encode time reduction on speed 1, compared to corner matching with AVX2.

Change-Id: Ib8fb933a595c7e36d046ef0fd599bcf2cd99839d