Overhaul disflow algorithm
This patch applies a huge amount of bugfixes and optimizations to the
disflow algorithm in the flow estimation library. The new version of
this code generates better results than the current corner matching
code, while being many times faster.

This patch also includes an SSE4.1 implementation of
aom_compute_flow_at_point, which contains the hottest parts
of the algorithm.

This code is not enabled yet, so there is no BDRATE or encode speed
impact yet. But preliminary results suggest that disflow will give
slightly better results than corner matching, with a 2 - 2.5% reduction
in overall encode time (at speed 1)

Change-Id: Ib8fb933a595c7e36d046ef0fd599bcf2cd99839d