Fix unsigned integer overflow in sse computation
The functions, encoder_variance and encoder_highbd_8_variance,
compute the sse value for the area beyond that of width and
height aligned to 16. As the sse value computed in these
functions was stored in an unsigned integer format, this value
could exceed the unsigned integer limit. In this CL, the possible
data overflow is avoided by changing the datatype of the relevant
variables to uint64_t.

Change-Id: I64cbb31aead6d23c1e0112400581a8a0392186ed