Modify av1_foreach_transformed_block_in_plane()
In this CL, av1_foreach_transformed_block_in_plane() is
modified to separately handle the case when block size is equal
to transform block size. In this case, the visitor function is
invoked directly with appropriate arguments. An assert is added
to ensure that the visitor function is called at least once if
av1_foreach_transformed_block_in_plane() is called. This is a
bit-exact change.

For AVIF still-image encode,

             Encode Time
cpu-used     Reduction(%)
    6           0.309
    7           0.822
    8           0.643
    9           1.087

Change-Id: Ia7b7398c9b13b2a50554612c4c7417fcc9e826bb