Improve frame interpolation filter resetting
Currently, the frame interp filter is allowed to be reset only for
EIGHTTAP_REGULAR filter type, which is overly strict. This CL
loosens the condition and allows the resetting following the defines
in av1_is_interp_needed(). This change will help some RTC encoding,
which will be implemented in the following CL.

STATS_CHANGED

Change-Id: I34e7e0faffcb80344468294c0b59cc03e1e6d58c