Improve frame interpolation filter resetting
In current code base, the frame interpolation filter is allowed to be
reset only for EIGHTTAP_REGULAR filter type. This condition is overly
strict and would miss opportunities for bit saving. This CL modifies
the way that the filter type counts are recorded to ensure the correct
frame filter type resetting decision, and also loosens the condition
for frame filter resetting. This change will help some RTC encoding,
which will be implemented in the following CL.

STATS_CHANGED

Change-Id: I34e7e0faffcb80344468294c0b59cc03e1e6d58c