rtc-screen: Fixes to screen in nonrd pickmode
Add condition fo palette mode to be checked
if best mode is inter and has motion above some
threshold. Currently it is only be checked
if best mode is intra.

This improves bdrate and visual quality on
screen clips.

Also fix to make IDTX work if the
reuse_inter_prediction is 1 or 0.

Change-Id: Ia2809f8e40a4ae5c4fc0da8339da1fce0348c2c3