rtc-screen: Remove speed feature from speed 10
This speed feature is more aggresive in setting the
zeromv_skip (based on kVeryLowSad instead of kZeroSad),
which can cause artifacts.

This is causing artifacts, as observed in CRD (comment#4)
of the bug below. This CL resolves the visual artifacts
reported in the bug.

Bdrate change for speed 10 on rtc_screen:
avg_psnr/ovr_psnr/ssim, IC: 0.25/0.68/0.52, -5%.
Most clips are not affected, except for 2
(screen_recording and colwinscroll).

Bug: b/266687023

Change-Id: I8ab6b56519a145d2d603815a35eccbb00a7a6286