thread_loop: don't lock around execute() call
Right now the thread_loop() function for worker threads locks
worker->impl_->mutex_ around the execute(worker) call. This is
apparently intended to prevent the main thread from changing
worker->status_ while execute(worker) is running. But if the
change_state() sees worker->status_ is WORK, it will wait until
worker->status_ becomes OK. So thread_loop() does not need to lock
worker->impl_->mutex_ around the execute(worker) call.

When ThreadSanitizer reports a data race related to worker threads, it
includes worker->impl_->mutex_ among the mutexes held by the worker
thread when the data race occurred. This is misleading because that
mutex is unlikely to be relevant to the data race.

Change-Id: Ie3892fc32884fc51639cda0e75ac1bbec7f8e1cf