Add or clean up comments related to denoising
aom/aom_encoder.h: Document that aom_codec_encode() modifies the samples
in img->planes[i] if denoise-noise-level is nonzero, even though the img
input parameter is declared as a const pointer.

Use "param[in,out]" instead of "param[in/out]" in Doxygen comments.
https://www.doxygen.nl/manual/commands.html#cmdparam says:
  When a parameter is both input and output, [in,out] is used as
  attribute.

Document that the `sd` parameter of av1_receive_raw_frame() is both an
input and output.

The apply_denoise parameter of aom_denoise_and_model_run() should be
documented as an input.

Add a TODO comment to aom_denoise_and_model_run() because it assumes
sd->u_buffer and sd->v_buffer are not null pointers, so it doesn't
handle monochrome frames correctly.

Change-Id: Ie6eef41d7416e73a8d823cecabc0f9a840b220b7