Fix valgrind failure with parameterized test
When gtest registers parameterized tests, it formats the parameter using
PrintToString. The default implementation (PrintBytesInObjectToImpl)
prints a hex dump of all the bytes in the object. If the parameter is
a struct with any padding in it, valgrind reports this as
"Use of uninitialised value".

This is a known issue (https://github.com/google/googletest/issues/3805)
and was happening with ZeroMotionPropagationTestParams.

Fortunately there's a simple workaround: define an overload of PrintTo,
which PrintToString will use instead of calling PrintBytesInObjectToImpl.

Change-Id: I9fa6cb9127453077c8247fab7b0658c85d45e586