Fix test failure in disflow code
The CodingPathSyncLarge.SearchForHbdLbdMismatchLarge test constructs
an artificial input consisting of a square wave like pattern of pixels.
This creates a situation where the pixel gradients are all aligned in
the same direction across the whole image. This causes the flow matrix
in the disflow code to be ill-conditioned, resulting in a division by 0.

This patch addresses this issue by simply not refining any points which
have a non-invertible flow matrix. As the problematic case is rare, this
only results in minor changes to a few videos.

BDRATE impact (65 frames, speed 4):

lowres2: -0.004%
midres2: +0.004%
hdres2: 0.000%

STATS_CHANGED

Bug: aomedia:3385
Change-Id: Ib78348624559da4ff759368d6caa11280eef8ee6