Regularize flow matrix in disflow
Apply a weak regularization to the flow matrix in the disflow algorithm.
This has two main effects: The sizes of the steps taken are reduced
very slightly, and more importantly, it guarantees that the determinant
of the flow matrix is always > 0. Thus we can remove the special case
logic which handled the case where det M = 0.

This also provides a more complete fix to bug #3385, compared to
the earlier fix in commit 14bd46b:

Without regularization, the determinant of the flow matrix is 0
if and only if all of the pixel gradients in the patch are pointing
in the same direction. This can happen if (for example) the image
contains a pattern of parallel lines.

In this case, we cannot know for sure where the source patch should
be mapped to, since shifting along the lines will not affect the
prediction error. So, in the previous fix, we chose to not do anything.
But with the addition of a weak regularization term, we can at least
move to some point where the images match, even if it isn't unique.

BDRATE and encode time impact relative to the previous fix
are both below measurement noise level.

STATS_CHANGED

BUG: aomedia:3385
Change-Id: Ib34d920ad107205ee3657447fc0c1b41fa2cb14d