Fix MV clamping in av1_mv_pred
Due to a quirk of the AV1 specification, when we derive a candidate
motion vector from global motion, it is *not* clamped, in contrast
to all other motion vector candidates. This can lead to candidate MVs
which point far outside of the reference frame in certain cases.

Most of the code handles such motion vectors by applying extra
clamping, in a way which is consistent with the specification. However,
the function av1_mv_pred() did not do this, leading to an out-of-bounds
read, as seen in the linked bug report.

This patch fixes the bug by adding an explicit clamp operation in
av1_mv_pred(), which is calculated to avoid changing the prediction
result while also staying entirely within the frame's allocated
border region.

BUG: aomedia:3389
Change-Id: I15feb47793534629bbb97bd6fcd6b88a40528225