set_fwd_txfm_non_scale_range: suppress -Wstringop-overflow
with gcc 12.2.0 (noticed when doing a build targeting aarch64).

To control struct size, stage_range_col[] and stage_range_row[] are
int8_t (char) which triggers the warning.

For consistency, rather than disabling the warning around this function,
add an additional check to the for loop as was done in:
ce3ad28696 Calculate the txfm stage range according to bd

A follow up could remove the check and use a GCC pragma to silence the
warnings (-Wstringop-overflow & -Warray-bounds).

Change-Id: Ia9dc4550d7f005d05ed54c2442a4f89a4e92a706