Fix the assertion failure in convolve functions
An assertion failure in av1_convolve_2d_sr_c() was reported, and can be
reproduced while building a c-only debug build of aomenc. The crash
error message is as follows.
av1_convolve_2d_sr_c:
Assertion `0 <= sum && sum < (1 << (bd + FILTER_BITS + 1))' failed.

The cause of this failure is due to some filter coefficients in 12-tap
filter. In extreme case, the pixel value (8-bit) can be 0 or 255. Then,
the sum can be < 0 or >  (1 << (bd + FILTER_BITS + 1)) ,which causes
the assertion failure. This CL fixed the issue.

Bug: aomedia:3393
Bug: b:270190269
Change-Id: I0aba0ffab0fe080056af012daf7591d9422d8a1a