rtc: Fix to setting color sensitivity in nonrd
If the superblock level color sensitivity flag is
set for golden or altref, then force it on for the
mode selection (incorporate color into rd cost).
Only do it for blocks > 16x16 to avoid slowdown.

This reduces the occurrence of the block color artifacts
in the issue below.

Stats changed:

Bug: aomedia:3370
Change-Id: Iaeb989ba83b876560a03dd7c74be4eeae1340856