rtc: Color sensitivity for altref in nonrd
Add superblock-level color sensitivity flag for altref,
similar to golden, and use this in the nonrd pickmode.
If this color flag is set for altref: skip altref as
single reference mode (in get_ref_frame_use_mask(), as
is done for golden), and skip the compound mode (and add
the same there condition for golden reference).

The skip logic allows for some speedup with less chance
of color block artifacts. The logic is also conditoned on
source variance, bsize, and content_state, to minimize
regression on the test sets.

This reduces the occurrence of wrong color blocks, observed
in the issue below.

Stats changed:

Bug: aomedia:3370
Change-Id: Iaeb989ba83b876560a03dd7c74be4eeae1340856