rtc: Color sensitivity for altref in nonrd
Add superblock level color sensitivity flag for altref,
similar to golden, and use this in the pickmode.
If color is set for altref: skip both the single reference
and compound mode (for compound mode also add the condition
for golden if it's the second reference and has color set).

This reduces the occurrence of the block color artifacts
in the issue below.

Stats changed:

Bug: aomedia:3370
Change-Id: Iaeb989ba83b876560a03dd7c74be4eeae1340856