rtc: Color sensitivity for compound in nonrd
For RTC: add check to skip compound if there is
color detected for the second references used
for single layer nonrd (golden or altref).

For this we add superblock level color sensitivity
flag for altref, similar to golden, and this is
then used in the pickmode.

This reduces the occurrence of the artifacts in
the issue below.

Stats changed:

Bug: aomedia:3370
Change-Id: Iaeb989ba83b876560a03dd7c74be4eeae1340856