rtc: Color sensitivity for altref in nonrd
This is reland of:
https://aomedia-review.googlesource.com/c/aom/+/171201
with some adjustment of thresholds.

Add superblock-level color sensitivity flag for altref,
similar to golden, and use this in the nonrd pickmode.
If this color flag is set for altref: skip altref as
single reference mode (in get_ref_frame_use_mask(), as
is done for golden), and skip the compound mode (and add
the same there condition for golden reference).

Skip logic is targeted to small source variance (flat areas)
in y channel, as that is where the wrong color artifact
likely occurs. And thresholds for setting golden/alt
color sensitivity are based on resolution and content mode.

This reduces the occurrence of wrong color blocks, observed
in the issue below. Though the issue/artifact can still arise,
this change is fairly conservative to avoid regression on
the test sets.

Stats changed: avg/ovr/ssim, IC (speedup)
rtc_derf:
speed 7:  -0.022/-0.032/0.022,  -0.061
speed 8:  -0.157/-0.185/-0.140, -0.169
speed 9:   0.119/0.081/0.141,    0.014
speed 10: -0.091/-0.105/-0.132, -0.034

rtc:
speed 7:  0.227/0.196/0.354,  0.146
speed 8:  0.351/0.330/0.568,  0.219
speed 9:  0.126/0.138/0.433, -0.276
speed 10: 0.120/0.138/0.381, -0.389

rtc_screen:
speed 10: -0.371/-0.198/-0.408, 0.033

rtc-1080p:
speed 7:  -1.091/-0.835/-0.023, -0.263
speed 10: -0.382/-0.362/-0.193, -0.554

Bug: aomedia:3370
Change-Id: I2a3a2942b9ae7f27a66e66bd078a67a8f4cd3ee6