rtc: Color sensitivity for altref in nonrd
This is reland of:
https://aomedia-review.googlesource.com/c/aom/+/171201
with some adjustment of thresholds.

Add superblock-level color sensitivity flag for altref,
similar to golden, and use this in the nonrd pickmode.
If this color flag is set for altref: skip altref as
single reference mode (in get_ref_frame_use_mask(), as
is done for golden), and skip the compound mode (and add
the same there condition for golden reference).

Skip logic is targeted to small source variance (flat areas)
in y channel, as that is where the wrong color artifact
likely occurs. And thresholds for setting golden/alt
color sensitivity are based on resolution and content mode.

This reduces the occurrence of wrong color blocks, observed
in the issue below. Though the issue/artifact can still arise,
this change is fairly conservative to avoid regression on
the test sets.

Stats changed: avg/ovr/ssim, IC (speedup)
rtc_derf:
speed 7:  0.176/0.119/0.279,   0.263
speed 8: -0.029/-0.027/0.038, -0.042
speed 9:  0.159/0.134/0.198,   0.088
speed 10: 0.01/-0.001/0.011,  -0.062

rtc:
speed 8:  0.427/0.389/0.684,  0.318
speed 9:  0.154/0.137/0.521, -0.178
speed 10: 0.358/0.327/0.720, -0.426

rtc_screen:
speed 10: -0.045/-0.025/-0.095, -0.1

rtc-1080p:
speed 7:
speed 10:

Change-Id: I2a3a2942b9ae7f27a66e66bd078a67a8f4cd3ee6