Remove leftover homography-related code
During the development of AV1, we considered warp models with up to
8 parameters, which could implement an arbitrary perspective
projection, also called a homography.

Eventually we settled on only using up to 6-parameter models, which
implement affine transformations. But some of the code related to
the 8-parameter models was left in libaom. This isn't needed any more,
so remove it to help make the code a bit cleaner.

Change-Id: I7965b9ea5661488f27b87c3cebb039d04760a2c5