rtc-screen: Fix usage of blk_skip in nonrd
In the current code, for inter-modes, blk_skip
array was defined for 8x8 blocks (since tx_size
is >= 8X8 for nonrd) to reduce computations on
the memset and memcpy.

But for screen, tx_size = 4x4 may be used and so
to be correct/consistent we need to define blk_skip
for 4x4 blocks.

This reduces the crumb artifact in issue:
https://b.corp.google.com/issues/266687023

Change-Id: I3e6205d79d4e9debf8d58b0c52b8ccc2eea6d609