svc_encoder_rtc: Don't include aom_ports/bitops.h
aom_ports/bitops.h is not a public header of libaom.

Replace get_msb() with a simple portable implementation and name the
function ilog2(), for integer logarithm base 2. This algorithm is half
of Figure 5-14 in Hacker's Delight.

Also replace a die() call with fatal() because the usage help message
doesn't need to be printed there.

Bug: aomedia:3417
Change-Id: Id67fb1b2e72dbdb8f66ddb10406c2a96f3938284