Skip AV1E_SET_TILE_COLUMNS if value is the default
The default value of AV1E_SET_TILE_COLUMNS is 0. Skip the
aom_codec_control(AV1E_SET_TILE_COLUMNS) call if the value we want to
set it to is 0, because the update_extra_cfg() call after setting the
value is not that cheap.

Also change cast the return value of log2() to unsigned int because that
is the parameter type of AV1E_SET_TILE_COLUMNS.

Change-Id: Ie7d066865b12b8766f4d4798a914f1c573f60ba2