Refactor Neon dc predictors
Rather than having all possible logic for each square predictor in one
big function, extract the individual parts out into a few helpers and
call those from each concrete dc predictor implementation. This has the
nice advantage of significantly reducing the number of if-statements
needed, which was ocassionally costly if the compiler failed to inline
the original helpers.

This also has the advantage of giving us some nice helper functions that
we can reuse for the rectangular dc predictor implementations (which are
not yet implemented for Neon).

Change-Id: I337f0fd160b48fe9650878fbb75e0f6acd48282c