aomdec.c: Verify we can scale a frame
Before scaling a frame, verify we can scale it, i.e., the frame's format
must be 4:2:0 and the frame's bit depth must be the same as the first
frame's bit depth.

Move the image_format_to_string() function from aomenc.c to
common/tools_common.c so that it can be called by aomdec.c.

Bug: aomedia:3424
Bug: aomedia:3425
Change-Id: I0b0348e902cb85e6bb1db5add8db1db27de22aa6